CRASH: 'plague' affect damage causes 'invis' affect to be deleted.
As it happens inside the same loop, next attempt to access 'invis' affect will work with deallocated memory.
To fix this situation, affects are not deleted but kept in an extract list,
similar to characets and items. Extracted affects are ignored in affect_remove but can be safely iterated over.